Quarkus integration PoC
Trying to provide minimal extension functions as I gave up on developing a “real” Quarkus extension after numerous attempts and days of failures. I leave it to others to develop with undocumented extension frameworks and bytecode-magic.

Also, the Quarkus extension should be a standalone project with its own repository, as it either pollutes the multiplatform Gradle build or even has to be built with Maven due to lack of support.